refactor(core): Move createDeferredPromise into @n8n/utils (no-changelog)
Move `createDeferredPromise` and `IDeferredPromise` from `n8n-workflow`
into `@n8n/utils/promise/deferred-promise`, following the Phase 1 pattern.

`packages/workflow/src/deferred-promise.ts` becomes a thin re-export so
`n8n-workflow` keeps exporting both symbols for external consumers, while
all internal import sites now point at `@n8n/utils`.
