+++ packages/@n8n/utils/src/promise/deferred-promise.test.ts
@@ -1,4 +1,4 @@
-import { createDeferredPromise } from '../src/deferred-promise';
+import { createDeferredPromise } from './deferred-promise';
 
 describe('DeferredPromise', () => {
 	it('should resolve the promise with the correct value', async () => {
