+++ packages/@n8n/utils/src/promise/deferred-promise.ts
@@ -1,0 +1,17 @@
+type ResolveFn<T> = (result: T | PromiseLike<T>) => void;
+type RejectFn = (error: Error) => void;
+
+export interface IDeferredPromise<T> {
+	promise: Promise<T>;
+	resolve: ResolveFn<T>;
+	reject: RejectFn;
+}
+
+export function createDeferredPromise<T = void>(): IDeferredPromise<T> {
+	const deferred: Partial<IDeferredPromise<T>> = {};
+	deferred.promise = new Promise<T>((resolve, reject) => {
+		deferred.resolve = resolve;
+		deferred.reject = reject;
+	});
+	return deferred as IDeferredPromise<T>;
+}
