+++ packages/cli/src/workflows/triggers/trigger-execution-context.factory.ts
@@ -1,5 +1,7 @@
 import { Logger } from '@n8n/backend-common';
+import type { IWorkflowDb } from '@n8n/db';
 import { Service } from '@n8n/di';
+import type { IDeferredPromise } from '@n8n/utils/promise/deferred-promise';
 import {
 	ErrorReporter,
 	PollContext,
@@ -8,11 +10,8 @@ import {
 	type IGetExecutePollFunctions,
 	type IGetExecuteTriggerFunctions,
 } from 'n8n-core';
-
-import { ActiveExecutions } from '@/active-executions';
 import type {
 	ExecutionError,
-	IDeferredPromise,
 	IExecuteResponsePromiseData,
 	INode,
 	INodeExecutionData,
@@ -24,6 +23,7 @@ import type {
 } from 'n8n-workflow';
 import { Workflow, UnexpectedError, createRunExecutionData } from 'n8n-workflow';
 
+import { ActiveExecutions } from '@/active-executions';
 import { DuplicateExecutionError } from '@/errors/duplicate-execution.error';
 import { EventService } from '@/events/event.service';
 import { executeErrorWorkflow } from '@/execution-lifecycle/execute-error-workflow';
@@ -31,7 +31,6 @@ import { ExecutionService } from '@/executions/execution.service';
 import { WorkflowExecutionService } from '@/workflows/workflow-execution.service';
 import { WorkflowPublishedDataService } from '@/workflows/workflow-published-data.service';
 import { WorkflowStaticDataService } from '@/workflows/workflow-static-data.service';
-import type { IWorkflowDb } from '@n8n/db';
 
 export type TriggerFailureHandler = (opts: {
 	error: Error;
