+++ packages/workflow/src/deferred-promise.ts
@@ -1,17 +1,3 @@
-type ResolveFn<T> = (result: T | PromiseLike<T>) => void;
-type RejectFn = (error: Error) => void;
-
-export interface IDeferredPromise<T> {
-	promise: Promise<T>;
-	resolve: ResolveFn<T>;
-	reject: RejectFn;
-}
-
-export function createDeferredPromise<T = void>(): IDeferredPromise<T> {
-	const deferred: Partial<IDeferredPromise<T>> = {};
-	deferred.promise = new Promise<T>((resolve, reject) => {
-		deferred.resolve = resolve;
-		deferred.reject = reject;
-	});
-	return deferred as IDeferredPromise<T>;
-}
+// Thin re-export kept for backwards compatibility with external consumers of
+// `n8n-workflow`. The implementation now lives in `@n8n/utils`.
+export * from '@n8n/utils/promise/deferred-promise';
